+++ src/main/java/com/braintreegateway/graphql/enums/Recommendations.java
@@ -1,0 +1,8 @@
+package com.braintreegateway.graphql.enums;
+
+/**
+ * Represents available recommendation categories that can be retrieved using a PayPal customer session.
+ */
+public enum Recommendations {
+  PAYMENT_RECOMMENDATIONS
+}