+++ src/main/java/com/braintreegateway/graphql/enums/RecommendedPaymentOption.java
@@ -1,0 +1,9 @@
+package com.braintreegateway.graphql.enums;
+
+/**
+ * Represents available payment options related to PayPal customer session recommendations.
+ */
+public enum RecommendedPaymentOption {
+  PAYPAL,
+  VENMO
+}